refactor(lotto) : Divide the package domain and views
