+++ src/main/kotlin/lotto/domain/Lotto.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 class Lotto(val prizedAmount: Int) {
     var count: Int = 0
