+++ src/test/kotlin/lotto/LottoGameTest.kt
@@ -1,5 +1,7 @@
 package lotto
 
+import lotto.domain.COUNT_OF_NUMBERS
+import lotto.domain.LottoGame
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatThrownBy
 import org.junit.jupiter.api.DisplayName
